feat(trx): WIP implementing discount in trx

+++ src/main/java/com/dti/ecim/discount/dto/CreateGlobalDiscountRequestDto.java
@@ -18,7 +18,7 @@ public class CreateGlobalDiscountRequestDto {
     @NotBlank
     private String description;
 
-    private Long amountFlat;
+    private int amountFlat;
 
     private int amountPercent;
