+++ src/main/java/com/dti/ecim/discount/entity/Discount.java
@@ -34,7 +34,7 @@ public class Discount {
     @NotBlank
     private String description;
 
-    private Long amountFlat;
+    private int amountFlat;
 
     private int amountPercent;
